Split PE idle handling code to its own file, now in C++
