+++ src/conv-core/convcore.c
@@ -237,9 +237,8 @@ void* LrtsAlloc(int, int);
 void  LrtsFree(void*);
 #endif
 
-CpvStaticDeclare(int, cmiMyPeIdle);
 #if CMK_SMP && CMK_TASKQUEUE
-CsvDeclare(unsigned int, idleThreadsCnt);
+CsvExtern(unsigned int, idleThreadsCnt);
 CpvDeclare(void *, CsdTaskQueue);
 #endif
 
@@ -1733,43 +1732,9 @@ void (*handler)();
  *
  *****************************************************************************/
 
-void CsdBeginIdle(void)
-{
-  CcdCallBacks();
-#if CMK_TRACE_ENABLED && CMK_PROJECTOR
-  _LOG_E_PROC_IDLE(); 	/* projector */
-#endif
-#if CMK_SMP && CMK_TASKQUEUE
-  if (CpvAccess(cmiMyPeIdle) !=1) {
-    CpvAccess(cmiMyPeIdle) = 1;
-    CmiMemoryAtomicIncrement(CsvAccess(idleThreadsCnt));
-  }
-#else
-  CpvAccess(cmiMyPeIdle) = 1;
-#endif // CMK_SMP
-  CcdRaiseCondition(CcdPROCESSOR_BEGIN_IDLE) ;
-}
-
-void CsdStillIdle(void)
-{
-  CcdRaiseCondition(CcdPROCESSOR_STILL_IDLE);
-}
-
-void CsdEndIdle(void)
-{
-#if CMK_TRACE_ENABLED && CMK_PROJECTOR
-  _LOG_E_PROC_BUSY(); 	/* projector */
-#endif
-#if CMK_SMP && CMK_TASKQUEUE
-  if (CpvAccess(cmiMyPeIdle) != 0){
-    CpvAccess(cmiMyPeIdle) = 0;
-    CmiMemoryAtomicDecrement(CsvAccess(idleThreadsCnt));
-  }
-#else
-  CpvAccess(cmiMyPeIdle) = 0;
-#endif // CMK_SMP
-  CcdRaiseCondition(CcdPROCESSOR_BEGIN_BUSY) ;
-}
+void CsdBeginIdle(void);
+void CsdStillIdle(void);
+void CsdEndIdle(void);
 
 extern int _exitHandlerIdx;
